reallocationPlot

Visualization of predictions from CoDA model

Usage

reallocationPlot(
  dc_obj = c(),
  data = c(),
  comps = c(),
  comps_fup = NULL,
  comps.names = c(),
  outcome = c(),
  covs = c(),
  comparisons = c("one-v-one", "prop-realloc")[2],
  increase = c(),
  decrease = c(),
  xlab = c(),
  ylab = c(),
  xlim = c(),
  ylim = c(),
  ReallocationLimits = c(-60, 60),
  xaxis.by = 5,
  ribbon = TRUE,
  font = "Times New Roman",
  main = NULL,
  col = "black",
  alpha = 0.2,
  analysis_type = c("cross-sectional", "longitudinal")[1],
  total = NULL
)

Arguments

dc_obj

Object derived from lm_coda or lm_coda_long

comps

String with the name of the composition variables in dataf

comps_fup

String with the name of the follow-up composition variables in dataf (optional)

comps.names

Names of the composition variables to appear in the plot (it will use comps if not provided)

outcome

String with the name of the outcome in dataf

covs

String with the name of the covariates

comparisons

Two choices ("one-v-one" or "prop-realloc")

increase

What variable in the composition should increase in the visualization (replacement)

decrease

What variable/s in the composition should decrease in the visualization (replaced)

xlab

X axis label

ylab

Y axis label

xlim

Limits of the X axis (time reallocation limits)

ylim

Limits for the Y axis (autogenerated if not provided)

ReallocationLimits

limits for the reallocation time in minutes (vector of 2 numbers)

xaxis.by

X axis ticks at every xaxis.by minutes

ribbon

Logical. Whether to plot confidence intervals of the predictions

font

Font to use (default = "Times New Roman")

main

Title of the plot

col

Color for lines and confidence intervals

alpha

Alpha for transparency in confidence intervals

analysis_type

One of "cross-sectiona", "longitudinal", "change".

total

Total time to constrain the composition to.

Value

reallocation plot with predictions

Details

If follow-up composition is provided, it will run prospective and change models/visualizations